% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-by-step.R
\name{as_step}
\alias{as_step}
\title{Wrap a function to be used as a step}
\usage{
as_step(f, name = NULL)
}
\arguments{
\item{f}{A function to be wrapped.}

\item{name}{The name of the step. If \code{NULL}, the step does not proceed but
the function is applied.}
}
\value{
A function that takes a step-by-step object and additional
arguments, and returns the updated step-by-step object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
\code{as_step()} wraps a function to be used as a step in a step-by-step
process.
}
