% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-by-step.R
\name{end_step}
\alias{end_step}
\title{End a step-by-step process}
\usage{
end_step(object)
}
\arguments{
\item{object}{The object to end the step-by-step process for.}
}
\value{
The object with the step-by-step attributes removed.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
\code{end_step()} ends the step-by-step process and removes the step-by-step
attributes from the object.
}
