% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-by-step-verb.R
\name{next_step}
\alias{next_step}
\title{Take the next step}
\usage{
next_step(.object, .step, ...)
}
\arguments{
\item{.object}{A step-by-step object.}

\item{.step}{The step to take.}

\item{...}{Additional arguments to pass to the step function.}
}
\value{
The updated step-by-step object.
}
\description{
\code{next_step()} is used to take the next step in a step-by-step process.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
