% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-by-step-verb.R
\name{step-by-step-manipulation}
\alias{step-by-step-manipulation}
\alias{insert_step}
\alias{update_step}
\alias{delete_step}
\title{Manipulate steps}
\usage{
insert_step(
  object,
  fns,
  steps = names(fns),
  descriptions = NULL,
  before = NULL,
  after = NULL
)

update_step(object, step, f = NULL, description = NULL)

delete_step(object, step)
}
\arguments{
\item{object}{A step-by-step object.}

\item{fns}{A list of functions to be applied step by step.}

\item{steps}{A character vector of step names.}

\item{descriptions}{A character vector of step descriptions.}

\item{before, after}{The step before or after which to insert the new steps.}

\item{step}{The step to be updated or deleted.}

\item{f}{The function to be applied to the step.}

\item{description}{The description of the step.}
}
\value{
The updated step-by-step object.
}
\description{
These functions are used to manipulate the steps of a step-by-step object.
}
