% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-by-step.R
\name{step_by_step}
\alias{step_by_step}
\title{Create a step-by-step object}
\usage{
step_by_step(steps)
}
\arguments{
\item{steps}{A named vector of steps to be completed. The names of the
vector are the names of the steps, and the values are the descriptions of
the steps.}
}
\value{
A function that takes an object and returns a step-by-step object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
\code{step_by_step()} creates a step-by-step object that can be used to track the
progress of a process.
It is useful for long-running processes where you want to keep track of
the steps that have been completed and the steps that are still to be done.
}
