% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-by-step-verb.R
\name{wrap_step}
\alias{wrap_step}
\title{Wrap a function to be used as a step}
\usage{
wrap_step(f)
}
\arguments{
\item{f}{A function to be wrapped.}
}
\value{
A function that takes a step-by-step object and additional
arguments, and returns the updated step-by-step object.
}
\description{
\code{wrap_step()} wraps a function to be used as a step in a step-by-step
process.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
