% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-by-step.R
\name{step_by_step}
\alias{step_by_step}
\title{Wrap functions for step-by-step data processing}
\usage{
step_by_step(fns, descriptions = NULL, ..., steps = names(fns))
}
\arguments{
\item{fns}{A list of functions to be applied step by step.}

\item{descriptions}{A character vector of step descriptions.}

\item{...}{Additional arguments for attributes.}

\item{steps}{A character vector of step names.}
}
\value{
A function that takes \code{data} and returns an object that inherits from
\code{adverbial_object_step_by_step}.
}
\description{
\code{step_by_step()} creates a step-by-step object that allows you to apply
functions to data in a sequence.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
